requirements.txt
Django==3.2.3
Pillow==8.2.0
